made grid a separate component, separated components into dev and photo folders
